Consolidate NIST and SFTP parity tests into unit test suite and cleanup temporary files
